Rename threading module to prevent confusion with Python's built-in threading module.
